Fixed "Show All Items" flag in the shop items list and Edit Item
The items list function will now return items depending on whether the
user wanted purchased items to be returned or not.

Fixed a side effect of implementing edit item the way I have. As a
result, if an items' name is edited, it's categoryID should also be
set. This is required in order to insert the new name into the
shopitemscatalog otherwise we'll get a FK failure.