FEATURE: When a list is shared with a user all categories associated with items in the shared list are also shared. The user cannot edit or delete those categories. Added access right checks to these operations. System categories (with userID=1) are also displayed for each user. These can't be edited or deleted either.
NOTE:
This does create some problems in the sense that the factory categories which are created for each user will end up appearing as duplicates. Unfortunately, no clear solution is apparent at this time.
Sharing of categories is needed because otherwise the UI which renders then hierarchically will not render children (items) of missing categories.