Changed the anatomy of `shopitems_price` table. For a given day, it now maintains the running sum of price of items and the number of items recorded that day.
Added new class `ShopItemAndStats` which extends ShopItem and contains information about the items' statistics. This is now used to return info about a newly created or edited item. Client's that do not wish to read this extra information, do not have to change their code.

A few minor bug fixes in the edit item area.